airflow/requirements.txt
requests==2.32.4
kafka-python==2.2.15
